backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0

